+++ __tests__/index.ts
@@ -1,3 +1,4 @@
+export * as biguery from './bigquery'
 export * as cloudfunctions from './cloudfunctions'
 export * as cloudfunctionsv2 from './cloudfunctionsv2'
 export * as cloudrun from './cloudrun'