+++ src/bigquery/index.ts
@@ -1,0 +1,7 @@
+import { requireCmekDataset } from "./requireCmekDataset";
+import { requireCmekTable } from "./requireCmekTable";
+
+export const bigqueryPolicies = [
+	requireCmekDataset,
+	requireCmekTable,
+];